Fix test and move triggers
