+++ bluebottle/activities/triggers.py
@@ -1,3 +1,5 @@
+from bluebottle.activities.models import Activity
+from bluebottle.activities.states import ActivityStateMachine
 from bluebottle.fsm.effects import TransitionEffect
 from bluebottle.fsm.triggers import ModelChangedTrigger
 
@@ -10,5 +12,9 @@ class Complete(ModelChangedTrigger):
         "There are no errors or missing fields"
         return (
             not list(self.instance.errors) and
-            not list(self.instance.required)
+            not list(self.instance.required) and
+            self.instance.status != ActivityStateMachine.needs_work.value
         )
+
+
+Activity.triggers = [Complete]
